added permission check system for clear messages

+++ src/slash-commands/commands/clearMessages.js
@@ -3,6 +3,13 @@ const { Channel } = require('discord.js');
 const clearMessages = async (interaction, CMD_NAME, options) => {
 	const numberOfMessages = options.getNumber('number') || 1;
 
+	if (!interaction.memberPermissions.has('MANAGE_MESSAGES')) {
+		return interaction.reply({
+			content: 'HEY HEY HEY there, I see what you trynna do there :eyes:',
+			ephemeral: false,
+		});
+	}
+
 	const { size } = await interaction.channel.bulkDelete(
 		numberOfMessages,
 		true,
